requirements.txt
fido2
solo-python
pytest
pytest-ordering
